# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: 너와다음 Backend API
  version: "1.0.0"
  summary: 선택지로 이야기를 이끄는 AI 인터랙티브 스토리 플랫폼
  description: |
    본문은 매 턴 AI 가 생성한다. 플레이 중 자유입력은 없고 사용자 입력면은 `choiceId` 하나뿐이다 (I-1).

    - 서비스 전체 단일 등급 = **15세 이용가**. 작품별 등급이 없다 (R10.1).
    - 모든 AI 응답은 Safety L2 를 통과한 뒤에만 사용자에게 도달한다 (I-2).
    - Chapter 전환과 Ending 선언은 **서버가 GameState 로 판정**한다. AI 제안값은 참고만 한다 (I-10).

    **구현 진행 상태는 이 계약과 다를 수 있다.** 계약이 앞서고 구현이 따라온다.
    `x-implemented-by` 확장이 각 필드·연산의 담당 작업을 가리킨다.
  license:
    name: 비공개
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/oauth/{provider}:
    post:
      tags: [auth]
      operationId: loginWithOAuth
      summary: 소셜 로그인 · 가입
      description: |
        최초 로그인이면 계정을 만든다. 이때 `birthDate` 와 `consents[]` 가 함께 와야 한다.

        - `birthDate` 또는 `consents[]` 누락 → `400 CONSENT_REQUIRED`
        - 만 15세 미만 → `403 AGE_RESTRICTED` (R10.2). **계정을 만들지 않는다**
        - 만 나이는 KST 기준으로 계산한다

        **MVP 는 Google 만 구현한다** (§13-11). Apple 은 상위 문서에 있으나 범위 밖이다.
        이메일 가입·로그인(`/auth/email/*`)은 같은 정정으로 계약에서 제외했다.
      security: []
      x-implemented-by: B-12, B-13
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google]
          description: MVP 는 `google` 뿐이다 (§13-11).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthLoginRequest'
      responses:
        '200':
          description: |
            발급된 액세스 토큰. **리프레시 토큰은 본문에 없다** — `Set-Cookie` 로만 나간다
            (ADR-0008, 이슈 #278).
          headers:
            Set-Cookie:
              $ref: '#/components/headers/RefreshTokenCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/ConsentRequired'
        '403':
          $ref: '#/components/responses/AgeRestricted'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: 액세스 토큰 재발급
      description: |
        **요청에 본문이 없다** (ADR-0008, 이슈 #278). 자격 증명은 로그인·재발급이 구운
        `HttpOnly` 쿠키 하나이며, 브라우저가 이 경로에만 자동으로 붙인다.

        **이 경로 하나만 CSRF 토큰을 요구한다.** `/api/v1/**` 의 나머지는 면제 그대로다 —
        거기에는 브라우저가 자동으로 붙이는 자격 증명이 없기 때문이다.

        - 쿠키가 없거나 유효하지 않다 → `401 UNAUTHENTICATED`
        - CSRF 토큰이 없거나 쿠키와 다르다 → `403 FORBIDDEN`

        호출하는 쪽은 **자격 증명을 실어야 한다** (`credentials: 'include'`). 실지 않으면
        브라우저가 쿠키를 붙이지 않고, 응답의 `Set-Cookie` 도 버린다.
      security: []
      x-implemented-by: B-12
      parameters:
        - name: X-XSRF-TOKEN
          in: header
          required: true
          schema:
            type: string
          description: |
            `XSRF-TOKEN` 쿠키에서 읽은 값을 그대로 돌려보낸다 (double-submit). 서버는 모든
            `/api/v1/**` 응답에 그 쿠키를 갱신하므로 별도로 발급받는 경로가 없다.
      responses:
        '200':
          description: |
            재발급된 액세스 토큰. **리프레시 토큰은 본문에 없다** — 새 쿠키가 함께 나간다.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/RefreshTokenCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # ── §13.2 Library ───────────────────────────────────────────
...
  /api/v1/stories/{storyId}:
    get:
      tags: [story]
      operationId: getStoryDetail
      summary: 작품 상세 (화면 2.2)
      description: |
        `ageRating` 은 **상수**다. 작품별 등급 컬럼을 두지 않는다 (R10.1, I-19).

        `totalEndings` 는 `isSecret = false` 인 엔딩만 센다 (R7.11).
        UGC 는 `authorDisplayName`(닉네임)을 표시한다. `playerRef` 는 노출하지 않는다 (§13-7).

        **인증 없이 열린다** (§13-54, #306). 목록에서 들어온 사람이 로그인 없이 작품을 볼 수 있고,
        플레이하려 할 때 로그인한다.

        **익명 요청에서는 `mySession` 이 `null` 이다** — 세션이 없는 회원과 같은 모양이라
        클라이언트가 분기를 하나 더 갖지 않는다. `story` 와 `characters` 는 같다.

        **볼 수 없는 작품은 `404` 다 — 익명이든 아니든 같다** (I-8). 승인되지 않았거나 `private`
        인 작품은 **없는 것과 구분되지 않는다**. `403` 을 쓰지 않는 이유는 그것이 작품의 존재를
        알리기 때문이다.

        **인증 없이 열리므로 IP 기준 호출 한도가 걸린다** (S-8). 탐색 셋이 **같은 창을 쓴다**.
      security: []
      x-implemented-by: B-16
      parameters:
        - $ref: '#/components/parameters/StoryId'
      responses:
        '200':
          description: 작품 · 캐릭터 · 내 세션
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    delete:
      tags: [authoring]
      operationId: deleteStory
      summary: 작품 삭제 (작성자)
      description: |
        **작성자 본인만** 지운다 (I-8). 공식 작품(`authorType = official`)은 이 경로로 지워지지
        않는다 — 작성자가 없으므로 어떤 요청자와도 일치하지 않는다.

        **되돌릴 수 없다.** 복구 경로를 두지 않는다 — 사용자에게 '삭제'라고 말한 이상
        되돌릴 수 있게 두는 쪽이 거짓말이다. 지운 뒤 이 작품은 목록·상세·이어하기 어디에도
        없고, 작성자의 "내 작품"에서도 사라진다. 다시 내려면 원고를 **새로 제출**한다
        (그 제출은 새 작품이 된다).

        **진행 중 세션은 없어진 작품과 같은 답을 받는다** — 다음 턴 요청은
        `423 STORY_SUSPENDED`(읽기 전용)이고, `resume` 은 `storySuspended` 다. 이미 읽은 턴과
        히스토리는 그대로 열린다. 정지(`suspended`)와 구분해 알리지 않는다 — 이어갈 수 없다는
        사실은 같고, 구분하면 작성자의 처분이 독자에게 드러난다.

        **기록은 남는다** (§13-44 와 같은 근거). 세션·턴·GameState 스냅샷·엔딩 도달률과
        신고·검수 이력은 지워지지 않는다 — 그것은 플레이한 사람들과 검수의 기록이지
        작성자의 것이 아니다. 내부 상태는 `reviewStatus = deleted` 이며, 그 값은 어떤 응답에도
        실리지 않는다(지운 작품이 조회되지 않으므로).

        이미 지워진 작품을 다시 지우면 `404` 다. 없는 작품·남의 작품과 구분하지 않는다.
      x-implemented-by: '#290'
      parameters:
        - $ref: '#/components/parameters/StoryId'
      responses:
        '204':
          description: 삭제됨. 본문 없음
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── §13.4 Session ───────────────────────────────────────────
...
  /api/v1/stories/{storyId}/appeal:
    post:
      tags: [authoring]
      operationId: appealStorySuspension
      summary: 정지된 작품의 재검토 요청
      description: |
        **자동으로 내려간 것을 사람이 다시 보는 길이다** (#290, §13-59). 신고 누적으로
        `suspended` 가 된 작품은 사람이 판정한 결과가 아니라 **셈**의 결과이며 (R8.9, §13-41),
        그것에 대고 작성자가 말할 자리가 없으면 정지 화면의 "이의가 있으면 문의해 주세요"는
        안내가 아니라 방치다.

        **이 요청은 작품의 상태를 바꾸지 않는다.** 정지된 작품은 **이미** 인간 검수 큐에 있다
        (§13-41 — 큐가 `in_review` 와 `suspended` 를 함께 본다). 여기서 `reviewStatus` 를
        움직이면 작성자가 검수 결과를 되돌리는 길이 된다 (I-8). 바뀌는 것은 둘이다.

        - **기록** — 작성자가 다투었다는 사실과 그 사유
        - **신호** — 검수 큐의 `ReviewQueueItem.appealed` 가 `true` 가 된다

        **순서는 바꾸지 않는다.** 요청은 공짜이므로 앞세우면 줄을 사는 길이 되고, 아직
        아무도 보지 못한 작품들이 그만큼 밀린다.

        **정지 건마다 한 번이다.** 사람이 판정하면(통과든 반려든) 그 요청은 답을 받은 것이
        되고, 다시 정지되면 다시 요청할 수 있다. 답을 받기 전 두 번째 요청은 `409` 다.

        **사람이 반려한 작품(`rejected`)은 대상이 아니다.** 이미 사람이 본 결과이며, 고쳐서
        다시 내는 길이 그쪽에 있다 (§13-40).

        **본문을 되돌려주지 않는다.** 되돌려줄 것이 없다 — 상태는 그대로 `suspended` 이고
        그 값은 `GET /api/v1/me/stories` 가 이미 준다.
      x-implemented-by: B-57
      parameters:
        - $ref: '#/components/parameters/StoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppealRequest'
      responses:
        '202':
          description: 접수됨. 복구를 약속하지 않는다 — 판정은 사람이 한다
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            재검토를 요청할 수 있는 자리가 아니다.

            - `STORY_NOT_SUSPENDED` — 내려가지 않은 작품이다
            - `ALREADY_EXISTS` — 이번 정지에 대해 이미 요청했고 아직 답을 받지 않았다
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: STORY_NOT_SUSPENDED
                message: 지금은 재검토를 요청할 수 없어요.
                details: {}

  # ── §13.9 신고 ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '`Authorization: Bearer <accessToken>`'
  headers:
    RefreshTokenCookie:
      description: |
        리프레시 토큰 (ADR-0008, 이슈 #278). **이 헤더가 리프레시 토큰이 나가는 유일한 자리다.**

        속성 넷 중 셋을 서버가 고정한다.

        - `HttpOnly` — JS 가 읽지 못한다. 이 결정의 이유 자체다
        - `Path=/api/v1/auth/refresh` — 브라우저가 **다른 어떤 경로에도 붙이지 않는다**
        - `SameSite=Strict` — 교차 사이트 요청에 붙지 않는다

        `Secure` 만 배포 설정이며 배포된 모든 환경에서 켜진다. 수명은 리프레시 토큰의 수명과
        같다.

        **프론트가 이 값을 읽을 수 없고, 읽을 필요도 없다.** 재발급을 부를 때 브라우저가 붙인다.
      schema:
        type: string
        examples: ['nwd_rt=<token>; Path=/api/v1/auth/refresh; Max-Age=2592000; HttpOnly; SameSite=Strict']
  parameters:
...
    StoryId:
      name: storyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ConsentRequired:
      description: '`birthDate` 또는 `consents[]` 누락 (§13.1)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: CONSENT_REQUIRED
            message: 가입을 계속하려면 추가 정보가 필요해요.
            details: {}
    ValidationError:
      description: 입력 검증 실패. `details.fields` 에 필드별 사유가 온다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: VALIDATION_ERROR
            message: 입력값을 확인해 주세요.
            details:
              fields:
                turnNo: must be greater than or equal to 0
...
    Unauthenticated:
      description: 토큰 없음 또는 만료
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHENTICATED
            message: 로그인이 필요해요.
            details: {}
    AgeRestricted:
      description: 만 15세 미만 (R10.2). 계정을 만들지 않는다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: AGE_RESTRICTED
            message: 만 15세부터 이용할 수 있어요.
            details: {}
    Forbidden:
      description: 소유자가 아니거나 지금 수행할 수 없는 작업이다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: FORBIDDEN
            message: 이 작업을 수행할 권한이 없어요.
            details: {}
    NotFound:
      description: 리소스 없음. **어떤 리소스인지 밝히지 않는다** — 존재 여부 자체가 정보다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: 요청하신 정보를 찾을 수 없어요.
            details: {}
...
    TooManyRequests:
      description: |
        **429 는 셋으로 구분한다** (§13-14d). 하나로 합치지 않는다 — 클라이언트 처리가 다르다.

        - `RETRY_COOLDOWN` — 연속 실패 3회. `details.retryAfterSeconds` 동안 재시도 비활성
        - `RATE_LIMITED` — 분당 호출 한도 초과
        - `QUOTA_EXCEEDED` — 일일 토큰·생성 한도 초과
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            retryCooldown:
              value:
                error: RETRY_COOLDOWN
                message: 잠시 후 다시 시도해 주세요.
                details:
                  retryAfterSeconds: 30
            rateLimited:
              value:
                error: RATE_LIMITED
                message: 요청이 너무 빨라요. 잠시 후 다시 시도해 주세요.
                details: {}
            quotaExceeded:
              value:
                error: QUOTA_EXCEEDED
                message: 오늘 이용할 수 있는 양을 모두 사용했어요.
                details: {}
...
  schemas:
...
    ErrorCode:
      type: string
      description: |
        서버가 내보내는 모든 에러의 `error` 값. `com.neowadaeum.common.error.ErrorCode` 와 **정확히 일치**한다.
        스펙에 없는 코드를 만들지 않는다.
      enum:
        - CONSENT_REQUIRED
        - INVALID_CHOICE
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - AGE_RESTRICTED
        - STORY_LIMIT_REACHED
        - FORBIDDEN
        - NOT_FOUND
        - TURN_CONFLICT
        - CONCURRENT_GENERATION
        - SESSION_ALREADY_ACTIVE
        - ALREADY_EXISTS
        - REVIEW_NOT_PENDING
        - STORY_NOT_SUSPENDED
        - SAFETY_BLOCKED
        - STORY_SUSPENDED
        - RETRY_COOLDOWN
        - RATE_LIMITED
        - QUOTA_EXCEEDED
        - CONTEXT_BUDGET_EXCEEDED
        - INTERNAL_ERROR
        - PROVIDER_ERROR
        - GENERATION_TIMEOUT
    Error:
      type: object
      description: |
        **모든 에러가 이 형태 하나로 수렴한다.**

        스택트레이스·예외 클래스명·SQL·내부 경로·세이프티 차단 사유·거절된 입력 원문을 담지 않는다 (S-6, R9.6, S-3).
        클라이언트는 `error` 코드로 문구를 매핑한다. 서버 `message` 를 UI 에 그대로 쓴다고 가정하지 않는다.
      required: [error, message, details]
      properties:
        error:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 안전한 폴백 문구
        details:
          type: object
          additionalProperties: true
          description: 코드별 부가 정보. **`null` 이 되지 않는다** — 값이 없으면 빈 객체다
    AuthorType:
      type: string
      enum: [official, user]
      description: 공식 작품과 사용자 작품을 같은 섹션에 섞지 않는 근거 (R13.1)
...
    OAuthLoginRequest:
      type: object
      description: |
        `birthDate` / `consents` 는 **최초 로그인(가입)에만** 필요하다. 기존 회원은 `idToken` 만 보낸다.
        누락이면 `400 CONSENT_REQUIRED` 로 무엇이 더 필요한지 알린다.
      required: [idToken]
      properties:
        idToken:
          type: string
          description: Provider 가 발급한 ID 토큰. 서버가 검증한다
        birthDate:
          type: [string, 'null']
          format: date
          description: 만 나이 계산은 KST 기준이다. 원본을 저장한다(생일 경과 처리)
        consents:
          type: [array, 'null']
          items:
            $ref: '#/components/schemas/ConsentItem'
    ConsentItem:
      type: object
      required: [consentType, version, agreed]
      properties:
        consentType:
          type: string
          enum: [tos, privacy, ai_notice, age]
          description: |
            `age` 는 **보내도 저장되지 않는다** — 서버가 생년월일로 판정해 스스로 한 줄을
            남기며, 클라이언트가 보낸 것까지 저장하면 동의 이력에 같은 종류가 두 줄 남는다
            (R10.2, §13-24, #270). 거절하지는 않는다
        version:
          type: string
          description: |
            `GET /api/v1/consents` 가 알려 준 값을 **그대로 되돌려 보낸다.**
            클라이언트가 상수로 들고 있으면 개정된 날부터 옛 판본이 이력에 남는다 (#261)
        agreed:
          type: boolean
...
    TokenResponse:
      x-notice-exempt: '인증·계정 화면이다. AI 생성물을 그리지 않으므로 사전 고지의 대상이 아니다'
      type: object
      description: |
        `playerRef` 를 반환하지 않는다. 클라이언트가 알 필요가 없다.

        **`refreshToken` 도 반환하지 않는다** (ADR-0008, 이슈 #278). 본문으로 한 번 주면 JS 가
        읽었다는 뜻이고, 그러면 `HttpOnly` 가 주는 보장이 문장으로만 남는다. `RefreshRequest`
        가 계약에서 사라진 것도 같은 이유다 — 본문으로도 보낼 수 있는 우회로를 남기지 않는다.
      required: [accessToken, tokenType, expiresIn]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          description: 액세스 토큰 만료까지 남은 초

    # ── §13.2 Library ───────────────────────────────────────
...
    StoryDetailResponse:
      type: object
      required: [story, characters, mySession, noticeText]
      properties:
        story:
          $ref: '#/components/schemas/StoryDetail'
        characters:
          type: array
          description: 빈 배열이면 클라이언트가 섹션을 숨긴다
          items:
            $ref: '#/components/schemas/CharacterCard'
        mySession:
          oneOf:
            - $ref: '#/components/schemas/MySessionBrief'
            - type: 'null'
        noticeText:
          type: string
          description: |
            AI 사전 고지 문구. **하드코딩하지 않는다** — `service_config` 에서 온다 (R11.1).
            Footer 가 상시 표시하므로 이 응답이 함께 싣는다 — 화면마다 `/landing` 을 다시 부르지 않는다.
    StoryDetail:
      type: object
      required: [storyId, title, heroImage, genres, description, worldIntro, ageRating, authorType, authorDisplayName, totalChapters, totalEndings]
      properties:
        storyId:
          type: string
          format: uuid
        title:
          type: string
        heroImage:
          type: [string, 'null']
        genres:
          type: array
          items:
            type: string
        description:
          type: string
        worldIntro:
          type: string
        ageRating:
          type: string
          const: 15세 이용가
          description: '**상수다** (R10.1, I-19). 작품별 등급 컬럼을 두지 않는다'
        authorType:
          $ref: '#/components/schemas/AuthorType'
        authorDisplayName:
          type: [string, 'null']
          minLength: 2
          maxLength: 12
          examples: ['연우']
          description: |
            UGC 작성자 닉네임 (§13-7). **`playerRef` 를 노출하지 않는다.**

            **`@` 는 값에 포함되지 않는다.** 화면이 붙이는 표기이며(와이어프레임 3g 의 `@연우`),
            받은 값 앞에 `@` 를 그대로 붙이면 된다 — 서버는 `@` 로 시작하는 이름을 거절한다.

            2~12자이며 한글 · 영문 · 숫자 · `_` · `-` · 공백만 허용한다. 양끝 공백은 없고
            내부 연속 공백은 하나로 모인 상태로 온다 (#287).
        totalChapters:
          type: integer
        totalEndings:
          type: integer
          description: '`isSecret = false` 인 엔딩만 센다 (R7.11)'
    CharacterCard:
      type: object
      required: [characterId, name, role, portraitImage, oneLine]
      properties:
        characterId:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: [string, 'null']
        portraitImage:
          type: [string, 'null']
        oneLine:
          type: [string, 'null']
    MySessionBrief:
      type: object
      required: [sessionId, chapterNo, chapterTitle, lastSceneSummary]
      properties:
        sessionId:
          type: string
          format: uuid
        chapterNo:
          type: integer
        chapterTitle:
          type: [string, 'null']
        lastSceneSummary:
          type: [string, 'null']

    # ── §13.4 Session ───────────────────────────────────────
...
    AppealRequest:
      type: object
      description: |
        **사유는 필수다.** 정지된 작품은 이미 검수 큐에 있으므로 (§13-41), 사유 없는 요청은
        검수자에게 아무것도 주지 않는다 — 그러면 이 경로는 큐를 다시 채우는 버튼일 뿐이다.

        **이 문자열은 검수자만 읽는다** (S-11). 응답에도 검수 큐의 한 줄에도 실리지 않으며,
        AI 페이로드로도 가지 않는다.
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
...